requirements.txt
requests>=2.9.1
aiohttp
aiortc
opencv-python==4.8.1.78
opencv-contrib-python
pyshine
numpy
imutils
pyrogram
pysftp
yt-dlp
paho-mqtt
libtorrent
torrentp
matplotplib
pandas
